vyra-master/requirements.txt
 # --- FastAPI & Server ---
fastapi
uvicorn[standard]

# --- Environment ---
python-dotenv

# --- Data / DB ---
pydantic
pydantic-settings
sqlalchemy
alembic
psycopg2-binary

# --- Password & Auth ---
bcrypt>=4.0.0
passlib[bcrypt]
python-jose[cryptography]
python-multipart
ldap3

# --- LLM / Embeddings ---
google-generativeai
sentence-transformers
numpy

# --- PDF / Document Processing ---
pypdf
python-docx
openpyxl
xlrd
python-pptx
fpdf2>=2.7.0

# --- OCR (Taranmış PDF Desteği) ---
easyocr
pdf2image

# --- Networking ---
requests

# --- Machine Learning (CatBoost Hybrid) ---
catboost>=1.2.0
rapidfuzz>=3.0.0

# --- Utils ---
tqdm

# --- Security / Rate Limiting ---
slowapi>=0.1.9

# --- LDAP / Active Directory ---
ldap3>=2.9.0
cryptography>=41.0.0

# --- Testing ---
pytest>=9.0.0
pytest-cov>=7.0.0
